Use TARGET_ARCH to pass -mmcu=
TARGET_ARCH is a little-used and little-known built-in variable of
make that defaults to an empty value but was intended to specify
any cross-compiling architecture flags that may be needed.

It is also better to use TARGET_ARCH rather than CFLAGS or LDFLAGS
as the it is only used once in both LINK.c and COMPILE.c whereas
LINK.c references both CFLAGS and LDFLAGS so there would be pointless
multiple reference to -mmcu.

Signed-off-by: Zygmunt Krynicki <zkrynicki@gmail.com>